fix: build Transcendence on macOS - add missing source files and 64-bit pointer fixes
- Add 16bitBlt.cpp, 16bitImageDrawing.cpp, 8bitDrawCircle.cpp, 8bitImageDrawing.cpp to build
- Fix 64-bit pointer casts in 16bitBlt.cpp (uintptr_t instead of DWORD)
- Add inline to strPatternSubst in KernelString.h to fix duplicate symbol
- Add SDL platform files (CG16bitFontSDL, CSoundMgrSDL, CTimerRegistrySDL, GameUIBridge)
- Add CStringStub.cpp with minimal string/kernel stubs for macOS
- Add Windows API stubs in DirectXUtilCompat.h (BITMAP, GetObject, SetDIBitsToDevice, etc)
- Update CFileWriteStream.cpp, CMemoryStream.cpp, CObject.cpp, CSymbolTable.cpp with minimal implementations

+++ Alchemy/DirectXUtil/16bitBlt.cpp
@@ -643,12 +643,12 @@ void DrawBltShimmer (CG16bitImage &Dest,
 					}
 				else if (*pAlphaPos == 255 || *pDestPos == 0)
 					{
-					if (PERM((DWORD)pDestPos * LARGE_PRIME2 + dwRnd) < byIntensity)
+					if (PERM((DWORD)(uintptr_t)pDestPos * LARGE_PRIME2 + dwRnd) < byIntensity)
 						*pDestPos = *pSrcPos;
 					}
 				else
 					{
-					if (PERM((DWORD)pDestPos * LARGE_PRIME2 + dwRnd) < byIntensity)
+					if (PERM((DWORD)(uintptr_t)pDestPos * LARGE_PRIME2 + dwRnd) < byIntensity)
 						{
 						DWORD pxSource = *pSrcPos;
 						DWORD pxDest = *pDestPos;
@@ -738,7 +738,7 @@ void DrawBltShimmer (CG16bitImage &Dest,
 			while (pSrcPos < pSrcPosEnd)
 				{
 				if ((*pSrcPos != wSrcBackColor)
-						&& (PERM((DWORD)pDestPos * LARGE_PRIME2 + dwRnd) < byIntensity))
+						&& (PERM((DWORD)(uintptr_t)pDestPos * LARGE_PRIME2 + dwRnd) < byIntensity))
 					*pDestPos = *pSrcPos;
 
 				pDestPos++;
